Find minimal indent for glob-defined fragments
